Prepare internal code for switching IOSArray to accept jint instead of NSUInteger. 	Change on 2014/09/03 by kstanger <kstanger@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=74726070

+++ jre_emul/Classes/IOSArray.m
@@ -110,7 +110,7 @@ void IOSArray_throwOutOfBoundsWithMsg(jint size, jint index) {
   @throw AUTORELEASE([[JavaLangArrayIndexOutOfBoundsException alloc] initWithNSString:msg]);
 }
 
-- (NSString *)descriptionOfElementAtIndex:(NSUInteger)index {
+- (NSString *)descriptionOfElementAtIndex:(jint)index {
   [self doesNotRecognizeSelector:_cmd];
   return nil;
 }
